[channels,urbdrc] fix enum-enum conversion
* fix integer type of port, use uint8_t
* cast enum values to uint8_t (enum values are used as constants here
  and multiple different ones are mixed together which provokes this
  warning)

+++ channels/urbdrc/client/libusb/libusb_udevice.h
@@ -51,7 +51,7 @@ typedef struct
 	BYTE bus_number;
 	BYTE dev_number;
 	char path[17];
-	int port_number;
+	UINT8 port_number;
 	int isCompositeDevice;
 
 	LIBUSB_DEVICE_HANDLE* libusb_handle;
